+++ app/Http/Controllers/Admin/AdminController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+
+class AdminController extends Controller
+{
+    //
+}
