--- app/MyNew.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
-
-class MyNew extends Model
-{
-    protected $table = 'news';
-}
