+++ routes/web.php
@@ -34,51 +34,51 @@ Route::get('/logout', 'AuthController@logout')
 
 // АДМИНКА
 
-Route::group(['prefix' => 'administrator', 'middleware' => ['permission:admin_access']], function() {
+Route::get('administrator/', 'AuthController@showAdmin')
+    ->name('admin');
 
-    Route::get('/', 'AuthController@showAdmin')
-        ->name('admin');
+Route::group(['prefix' => 'administrator', 'namespace' => 'Admin', 'middleware' => ['permission:admin_access']], function() {
 
-    Route::get('/users', ['middleware' => ['ability:super_admin|user_admin,view_users'], 'uses' => 'UsersController@show'])
+    Route::get('/users', ['middleware' => ['ability:super_admin|user_admin,view_users'], 'uses' => 'UserController@show'])
         ->name('admin.users');
 
-    Route::post('/users/activate', ['middleware' => ['ability:super_admin|user_admin,activate_users'], 'uses' => 'UsersController@activate'])
+    Route::post('/users/activate', ['middleware' => ['ability:super_admin|user_admin,activate_users'], 'uses' => 'UserController@activate'])
         ->name('admin.users.activate');
 
-    Route::post('/users/delete', ['middleware' => ['ability:super_admin|user_admin,delete_users'], 'uses' => 'UsersController@delete'])
+    Route::post('/users/delete', ['middleware' => ['ability:super_admin|user_admin,delete_users'], 'uses' => 'UserController@delete'])
         ->name('admin.users.delete');
 
-    Route::get('/users/create', ['middleware' => ['ability:super_admin|user_admin,create_users'], 'uses' => 'UsersController@create'])
+    Route::get('/users/create', ['middleware' => ['ability:super_admin|user_admin,create_users'], 'uses' => 'UserController@create'])
         ->name('admin.users.create');
 
-    Route::post('/users/create', ['middleware' => ['ability:super_admin|user_admin,create_users'], 'uses' => 'UsersController@createPost'])
+    Route::post('/users/create', ['middleware' => ['ability:super_admin|user_admin,create_users'], 'uses' => 'UserController@createPost'])
         ->name('admin.users.create.post');
 
-    Route::get('/users/edit/{id}', ['middleware' => ['ability:super_admin|user_admin,edit_users'], 'uses' => 'UsersController@edit'])
+    Route::get('/users/edit/{id}', ['middleware' => ['ability:super_admin|user_admin,edit_users'], 'uses' => 'UserController@edit'])
         ->where('id', '[0-9]+')
         ->name('admin.users.edit');
 
-    Route::post('/users/edit/{id}', ['middleware' => ['ability:super_admin|user_admin,edit_users'], 'uses' => 'UsersController@editPost'])
+    Route::post('/users/edit/{id}', ['middleware' => ['ability:super_admin|user_admin,edit_users'], 'uses' => 'UserController@editPost'])
         ->where('id', '[0-9]+')
         ->name('admin.users.edit.post');
 
-    Route::get('/roles', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@show'])
+    Route::get('/roles', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@show'])
         ->name('admin.roles');
 
-    Route::post('/roles/delete', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@delete'])
+    Route::post('/roles/delete', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@delete'])
         ->name('admin.roles.delete');
 
-    Route::get('/roles/create', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@create'])
+    Route::get('/roles/create', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@create'])
         ->name('admin.roles.create');
 
-    Route::post('/roles/create', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@createPost'])
+    Route::post('/roles/create', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@createPost'])
         ->name('admin.roles.createPost');
 
-    Route::get('/roles/edit/{id}', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@edit'])
+    Route::get('/roles/edit/{id}', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@edit'])
         ->where('id', '[0-9]+')
         ->name('admin.roles.edit');
 
-    Route::post('/roles/edit/{id}', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RolesController@editPost'])
+    Route::post('/roles/edit/{id}', ['middleware' => ['ability:super_admin,create_edit_delete_roles'], 'uses' => 'RoleController@editPost'])
         ->where('id', '[0-9]+')
         ->name('admin.roles.editPost');
